Extract most of seeder to Gene
